Create Titanic Survival detector
The project aimed to predict the likelihood of survival for passengers on the Titanic Data Science  techniques in Python. The dataset was preprocessed to handle missing values and encode categorical variables. Different algorithms were trained on selected features such as  sex, age, gender, class, and embarked port. The Random Forest model  achieved the highest accuracy of [84.7]  among all the models on the test set, demonstrating its ability to predict survival outcomes effectively.